exploring the world -real api call
