+++ .babelrc
@@ -1,3 +1,3 @@
 {
-    "plugins": [["transform-remove-console", { "exclude": ["error", "warn"] }]]
+    // "plugins": [["transform-remove-console", { "exclude": ["error", "warn"] }]]
 }
