+++ index.css
@@ -42,3 +42,16 @@
 .res__cart>img{
     width: 100%;
 }
+
+.search__container{
+margin: 10px;
+}
+.search-input, button.search-btn {
+    font-size: 16px;
+    padding: 10px;
+    margin-right: 10px;
+}
+.search-btn{
+    background-color: rgb(71, 44, 44);
+    color: #fff;
+}
