+++ package.json
@@ -19,12 +19,12 @@
   "homepage": "https://github.com/pn50742/React-live#readme",
   "devDependencies": {
     "@babel/core": "^7.20.12",
-    "babel-plugin-transform-remove-console": "^6.9.4",
     "parcel": "^2.8.2",
     "process": "^0.11.10"
   },
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-loading-skeleton": "^3.1.0"
   }
 }
