--- src/Config.js
@@ -1,2071 +1,0 @@
-export const restrourentList = [
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "596129",
-        "name": "Chinese Wok",
-        "uuid": "9c7560af-0457-4e01-84cd-330d1b634889",
-        "city": "81",
-        "area": "Greenfields",
-        "totalRatingsString": "100+ ratings",
-        "cloudinaryImageId": "zptzllu39wk2ivjkvvrz",
-        "cuisines": [
-          "Chinese",
-          "Asian",
-          "Tibetan",
-          "Desserts"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 25000,
-        "costForTwoString": "₹250 FOR TWO",
-        "deliveryTime": 34,
-        "minDeliveryTime": 30,
-        "maxDeliveryTime": 40,
-        "slaString": "30-40 MINS",
-        "lastMileTravel": 4.900000095367432,
-        "slugs": {
-          "restaurant": "chinese-wok-(wok-express)-greenfields-greenfields",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "Ground Floor, SCO No- 705, Sector- B, Greenfield Colony, Faridabad, Faridabad, Faridabad, Haryana 121003",
-        "locality": "Greenfield Colony",
-        "parentId": 61955,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "60% off",
-          "shortDescriptionList": [
-            {
-              "meta": "60% off | Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "60% off up to ₹120 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "60% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "60% off up to ₹120 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "ribbon": [
-          {
-            "type": "PROMOTED"
-          }
-        ],
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "cid=5781030~p=1~eid=00000185-ee73-559a-119c-e33c00830164",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "4.9 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "596129",
-          "deliveryTime": 34,
-          "minDeliveryTime": 30,
-          "maxDeliveryTime": 40,
-          "lastMileTravel": 4.900000095367432,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": true,
-        "avgRating": "3.5",
-        "totalRatings": 100,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "22326",
-        "name": "Anupama Sweets and Restaurant",
-        "uuid": "0a900679-3b68-4a26-a06a-5ad7970cb340",
-        "city": "81",
-        "area": "Sector 34",
-        "totalRatingsString": "10000+ ratings",
-        "cloudinaryImageId": "skdojdzsqnxhohbrilqa",
-        "cuisines": [
-          "North Indian",
-          "Chinese",
-          "Desserts"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 25000,
-        "costForTwoString": "₹250 FOR TWO",
-        "deliveryTime": 31,
-        "minDeliveryTime": 31,
-        "maxDeliveryTime": 31,
-        "slaString": "31 MINS",
-        "lastMileTravel": 1.899999976158142,
-        "slugs": {
-          "restaurant": "anupama-sweets-and-restaurant-sector-34-greenfields_archive",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "SCF 3, Ashoka Enclave Part 1, Near Kanishka Tower, Sector 34, Faridabad",
-        "locality": "Near Kanishka Tower",
-        "parentId": 14538,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "50% off",
-          "shortDescriptionList": [
-            {
-              "meta": "50% off | Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "50% off up to ₹100 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "50% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "50% off up to ₹100 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "1.8 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "22326",
-          "deliveryTime": 31,
-          "minDeliveryTime": 31,
-          "maxDeliveryTime": 31,
-          "lastMileTravel": 1.899999976158142,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.8",
-        "totalRatings": 10000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "242283",
-        "name": "McDonald's",
-        "uuid": "5b534325-03c3-4592-a911-a1a72fd755c6",
-        "city": "81",
-        "area": "Sector 35",
-        "totalRatingsString": "5000+ ratings",
-        "cloudinaryImageId": "mejm5avxxzrmwyylklgt",
-        "cuisines": [
-          "American"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 40000,
-        "costForTwoString": "₹400 FOR TWO",
-        "deliveryTime": 42,
-        "minDeliveryTime": 40,
-        "maxDeliveryTime": 50,
-        "slaString": "40-50 MINS",
-        "lastMileTravel": 3.5,
-        "slugs": {
-          "restaurant": "mcdonalds-hr-faridabad-crown-interiorz-greenfields_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "HR Faridabad Crown Interior, Sarai Khawaja, 12/7, Delhi- Mathura Road Sector- 35 Faridabad-121001",
-        "locality": "HR Faridabad Crown Interiorz",
-        "parentId": 630,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "20% off",
-          "shortDescriptionList": [
-            {
-              "meta": "20% off | Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "20% off up to ₹50 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "20% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "20% off up to ₹50 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "3.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "242283",
-          "deliveryTime": 42,
-          "minDeliveryTime": 40,
-          "maxDeliveryTime": 50,
-          "lastMileTravel": 3.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "4.1",
-        "totalRatings": 5000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "52727",
-        "name": "Raina Di Hatti",
-        "uuid": "98961842-88d8-4dce-83da-c6a77c50712d",
-        "city": "81",
-        "area": "Sector 43",
-        "totalRatingsString": "10000+ ratings",
-        "cloudinaryImageId": "bz9ydlwnzbfl8dlp1uxf",
-        "cuisines": [
-          "North Indian",
-          "Chinese"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 20000,
-        "costForTwoString": "₹200 FOR TWO",
-        "deliveryTime": 32,
-        "minDeliveryTime": 30,
-        "maxDeliveryTime": 40,
-        "slaString": "30-40 MINS",
-        "lastMileTravel": 5,
-        "slugs": {
-          "restaurant": "raina-di-hatti-sector-37-greenfields_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "nr by omax hills sector 43 greenfields faridabad",
-        "locality": "Faridabad green fields",
-        "parentId": 15925,
-        "unserviceable": false,
-        "veg": true,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "20% off",
-          "shortDescriptionList": [
-            {
-              "meta": "20% off | Use PARTY",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "20% off | Use code PARTY",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "20% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use PARTY",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "20% off | Use code PARTY",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "ribbon": [
-          {
-            "type": "PROMOTED"
-          }
-        ],
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "cid=5770519~p=4~eid=00000185-ee73-559a-119c-e33d0083043d",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "52727",
-          "deliveryTime": 32,
-          "minDeliveryTime": 30,
-          "maxDeliveryTime": 40,
-          "lastMileTravel": 5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": true,
-        "avgRating": "4.1",
-        "totalRatings": 10000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "259065",
-        "name": "THE MAGICAL GRILL",
-        "uuid": "b6f108f9-8eaf-41c9-b980-d383ba9943c1",
-        "city": "81",
-        "area": "Sector 37",
-        "totalRatingsString": "500+ ratings",
-        "cloudinaryImageId": "ab2zsonzmmzqwnqt3bsa",
-        "cuisines": [
-          "North Indian",
-          "Chinese",
-          "Mughlai",
-          "Tandoor"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 25000,
-        "costForTwoString": "₹250 FOR TWO",
-        "deliveryTime": 31,
-        "minDeliveryTime": 31,
-        "maxDeliveryTime": 31,
-        "slaString": "31 MINS",
-        "lastMileTravel": 1.5,
-        "slugs": {
-          "restaurant": "the-magical-grill-greenfields_fbd-greenfields_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "Shop No 88 Main Huda Market Sector 37 Faridabad Haryana, 121003, India, Faridabad, Faridabad (Haryana), - 121003",
-        "locality": "Greenfields_FBD",
-        "parentId": 211160,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "FREE DELIVERY",
-          "shortDescriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "",
-          "shortDescriptionList": [
-            {
-              "meta": "Free Delivery",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "1.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "259065",
-          "deliveryTime": 31,
-          "minDeliveryTime": 31,
-          "maxDeliveryTime": 31,
-          "lastMileTravel": 1.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.8",
-        "totalRatings": 500,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "462153",
-        "name": "Bakingo",
-        "uuid": "20d1f0d2-e9af-4f80-a5cf-2148256648d6",
-        "city": "81",
-        "area": "Faridabad Sec-37",
-        "totalRatingsString": "500+ ratings",
-        "cloudinaryImageId": "ue5mktp2dj9z4j04sril",
-        "cuisines": [
-          "Bakery",
-          "Desserts",
-          "Beverages",
-          "Snacks"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 30000,
-        "costForTwoString": "₹300 FOR TWO",
-        "deliveryTime": 39,
-        "minDeliveryTime": 39,
-        "maxDeliveryTime": 39,
-        "slaString": "39 MINS",
-        "lastMileTravel": 1.7000000476837158,
-        "slugs": {
-          "restaurant": "bakingo-faridabad-greenfields",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "48-49 Plot No. 12/2 Sher Shah Suri Road Sector 37, Faridabad, Faridabad, Haryana-121003  (Back Side of 24X7 & Hype Gym)",
-        "locality": "Sher Shah Suri Road",
-        "parentId": 3818,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "40% off",
-          "shortDescriptionList": [
-            {
-              "meta": "40% off | Use SWIGGYIT",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "40% off up to ₹80 | Use code SWIGGYIT",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "40% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use SWIGGYIT",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "40% off up to ₹80 | Use code SWIGGYIT",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "1.7 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "462153",
-          "deliveryTime": 39,
-          "minDeliveryTime": 39,
-          "maxDeliveryTime": 39,
-          "lastMileTravel": 1.7000000476837158,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "4.3",
-        "totalRatings": 500,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "425076",
-        "name": "The Oriental Spice",
-        "uuid": "9a98cd55-58df-4cb3-b2da-1ba5e6048837",
-        "city": "81",
-        "area": "Sector 15",
-        "totalRatingsString": "500+ ratings",
-        "cloudinaryImageId": "bcqmo570zjg6sgxjc2m4",
-        "cuisines": [
-          "Chinese",
-          "Thai",
-          "Pan-Asian"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 29900,
-        "costForTwoString": "₹299 FOR TWO",
-        "deliveryTime": 48,
-        "minDeliveryTime": 48,
-        "maxDeliveryTime": 48,
-        "slaString": "48 MINS",
-        "lastMileTravel": 8.800000190734863,
-        "slugs": {
-          "restaurant": "the-oriental-spice-sector-15-sector-15",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "\"Abadi Deh/ Lal Dora, Khasra no 17, Sec-16A Near Hanuman Mandir, Daulatabad, Faridabad, Faridabad, Haryana, 121002 \"",
-        "locality": "Daulatabad",
-        "parentId": 11584,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "60% off",
-          "shortDescriptionList": [
-            {
-              "meta": "60% off | Use GUILTFREEJAN",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "60% off up to ₹150 | Use code GUILTFREEJAN",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "60% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use GUILTFREEJAN",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "60% off up to ₹150 | Use code GUILTFREEJAN",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "ribbon": [
-          {
-            "type": "PROMOTED"
-          }
-        ],
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            {
-              "name": "distance",
-              "fee": 7400,
-              "message": ""
-            },
-            {
-              "name": "time",
-              "fee": 0,
-              "message": ""
-            },
-            {
-              "name": "special",
-              "fee": 0,
-              "message": ""
-            }
-          ],
-          "totalFees": 7400,
-          "message": "",
-          "title": "Delivery Charge",
-          "amount": "7400",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "cid=5738834~p=7~eid=00000185-ee73-559a-119c-e33e00830725",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "8.8 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "425076",
-          "deliveryTime": 48,
-          "minDeliveryTime": 48,
-          "maxDeliveryTime": 48,
-          "lastMileTravel": 8.800000190734863,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": true,
-        "avgRating": "4.1",
-        "totalRatings": 500,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "548573",
-        "name": "TAMANG MOMOS & KATHI ROLL",
-        "uuid": "f38d6911-0e84-4fab-bc6e-3fd1a8dab8ad",
-        "city": "81",
-        "area": "Greenfields",
-        "totalRatingsString": "20+ ratings",
-        "cloudinaryImageId": "wzgpro6itsshiaohaaru",
-        "cuisines": [
-          "Chinese",
-          "Snacks",
-          "Fast Food",
-          "Italian"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 20000,
-        "costForTwoString": "₹200 FOR TWO",
-        "deliveryTime": 37,
-        "minDeliveryTime": 37,
-        "maxDeliveryTime": 37,
-        "slaString": "37 MINS",
-        "lastMileTravel": 0.8999999761581421,
-        "slugs": {
-          "restaurant": "tamang-momos-&-kathi-roll-greenfields-greenfields",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "D-BLOCK GALI NO. 4 NEAR TARUN NIKETAN PUBLIC SCHOOL SHIV COLONY SEHATPUR ROAD PALLA NO. 1 BADARPUR",
-        "locality": "Shiv Colony Sehatpur Road",
-        "parentId": 329727,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "FREE DELIVERY",
-          "shortDescriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "",
-          "shortDescriptionList": [
-            {
-              "meta": "Free Delivery",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "0.8 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "548573",
-          "deliveryTime": 37,
-          "minDeliveryTime": 37,
-          "maxDeliveryTime": 37,
-          "lastMileTravel": 0.8999999761581421,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.3",
-        "totalRatings": 20,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "45177",
-        "name": "Burger King",
-        "uuid": "a9a89711-468c-433a-934f-d19d4905aae3",
-        "city": "81",
-        "area": "Sector 35",
-        "totalRatingsString": "10000+ ratings",
-        "cloudinaryImageId": "slzuzut5irtkjmwtttip",
-        "cuisines": [
-          "Burgers",
-          "American"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 35000,
-        "costForTwoString": "₹350 FOR TWO",
-        "deliveryTime": 38,
-        "minDeliveryTime": 35,
-        "maxDeliveryTime": 45,
-        "slaString": "35-45 MINS",
-        "lastMileTravel": 3.5,
-        "slugs": {
-          "restaurant": "burger-king-crown-interiorz-mall-greenfields_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "Shop No SF 07, Second Floor, Crown Interiorz Mall, Faridabad",
-        "locality": "Crown Interiorz Mall",
-        "parentId": 166,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "60% off",
-          "shortDescriptionList": [
-            {
-              "meta": "60% off | Use STEALDEAL",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "60% off up to ₹120 | Use code STEALDEAL",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "60% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use STEALDEAL",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "60% off up to ₹120 | Use code STEALDEAL",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "3.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "45177",
-          "deliveryTime": 38,
-          "minDeliveryTime": 35,
-          "maxDeliveryTime": 45,
-          "lastMileTravel": 3.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "4.1",
-        "totalRatings": 10000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "22323",
-        "name": "Berco's -If you love Chinese",
-        "uuid": "716186ea-a94b-45a8-859a-50a7603bad9f",
-        "city": "81",
-        "area": "Sector 35",
-        "totalRatingsString": "1000+ ratings",
-        "cloudinaryImageId": "z7mijzhzmi0maigf8hwn",
-        "cuisines": [
-          "Chinese",
-          "Thai",
-          "Asian",
-          "Beverages",
-          "Desserts",
-          "Oriental",
-          "Seafood",
-          "Pan-Asian"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 50000,
-        "costForTwoString": "₹500 FOR TWO",
-        "deliveryTime": 37,
-        "minDeliveryTime": 35,
-        "maxDeliveryTime": 45,
-        "slaString": "35-45 MINS",
-        "lastMileTravel": 3.5,
-        "slugs": {
-          "restaurant": "bercos-sector-35-greenfields_archive",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "Shop R-2, 3rd Floor, Crown Interiorz Mall, Sector 35, Faridabad",
-        "locality": "Crown Interiorz Mall",
-        "parentId": 108,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "FREE DELIVERY",
-          "shortDescriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "",
-          "shortDescriptionList": [
-            {
-              "meta": "Free Delivery",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "ribbon": [
-          {
-            "type": "PROMOTED"
-          }
-        ],
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "cid=5656365~p=10~eid=00000185-ee73-559a-119c-e33f00830a5a",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "3.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "22323",
-          "deliveryTime": 37,
-          "minDeliveryTime": 35,
-          "maxDeliveryTime": 45,
-          "lastMileTravel": 3.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": true,
-        "avgRating": "4.0",
-        "totalRatings": 1000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "167097",
-        "name": "Kurries Aap ki Khatir",
-        "uuid": "fb51a55e-9551-4818-979d-8ca78ffa0284",
-        "city": "81",
-        "area": "Greenfields_FBD",
-        "totalRatingsString": "1000+ ratings",
-        "cloudinaryImageId": "uddvuyaf0evvn0mlnyzs",
-        "cuisines": [
-          "North Indian"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 20000,
-        "costForTwoString": "₹200 FOR TWO",
-        "deliveryTime": 37,
-        "minDeliveryTime": 35,
-        "maxDeliveryTime": 45,
-        "slaString": "35-45 MINS",
-        "lastMileTravel": 2.0999999046325684,
-        "slugs": {
-          "restaurant": "kurries-aap-ki-khatir-greenfields_archive",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "\t2nd 60 feet road molarband extn.badarpur border , District - South , STATE - Delhi",
-        "locality": "Greenfields",
-        "parentId": 121998,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "FREE DELIVERY",
-          "shortDescriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "",
-          "shortDescriptionList": [
-            {
-              "meta": "Free Delivery",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "2 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "167097",
-          "deliveryTime": 37,
-          "minDeliveryTime": 35,
-          "maxDeliveryTime": 45,
-          "lastMileTravel": 2.0999999046325684,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.3",
-        "totalRatings": 1000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "325116",
-        "name": "Masala Chaska",
-        "uuid": "37dc345e-840b-4b95-b913-bd35c5df478b",
-        "city": "81",
-        "area": "Sector 37",
-        "totalRatingsString": "100+ ratings",
-        "cloudinaryImageId": "o3gsgjdhsalz14rvujx4",
-        "cuisines": [
-          "North Indian",
-          "Chinese",
-          "Thalis"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 25000,
-        "costForTwoString": "₹250 FOR TWO",
-        "deliveryTime": 32,
-        "minDeliveryTime": 32,
-        "maxDeliveryTime": 32,
-        "slaString": "32 MINS",
-        "lastMileTravel": 1.5,
-        "slugs": {
-          "restaurant": "wah-ji-wah-greenfields_fbd-greenfields_fbd-2",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "shop no 76 huda market sec 37 faridabad",
-        "locality": "Greenfields_FBD",
-        "parentId": 132970,
-        "unserviceable": false,
-        "veg": true,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "40% off",
-          "shortDescriptionList": [
-            {
-              "meta": "40% off | Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "40% off up to ₹80 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "40% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "40% off up to ₹80 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "1.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "325116",
-          "deliveryTime": 32,
-          "minDeliveryTime": 32,
-          "maxDeliveryTime": 32,
-          "lastMileTravel": 1.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.7",
-        "totalRatings": 100,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "648823",
-        "name": "Biryani.Com",
-        "uuid": "4bba8c1b-d08d-449f-a78c-14fbad45c7d3",
-        "city": "81",
-        "area": "Sector 15_FBD",
-        "totalRatingsString": "50+ ratings",
-        "cloudinaryImageId": "63e1848761423bc61f76bcbf1802af4a",
-        "cuisines": [
-          "Biryani",
-          "Desserts",
-          "Beverages"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 50000,
-        "costForTwoString": "₹500 FOR TWO",
-        "deliveryTime": 40,
-        "minDeliveryTime": 40,
-        "maxDeliveryTime": 40,
-        "slaString": "40 MINS",
-        "lastMileTravel": 9.899999618530273,
-        "slugs": {
-          "restaurant": "biryani.com-sector-15_fbd-sector-15_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "17/3, Mathura Road, Sector 16A, Faridabad, Haryana 121002, India",
-        "locality": "Mathura Road",
-        "parentId": 12866,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "ribbon": [
-          {
-            "type": "PROMOTED"
-          }
-        ],
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            {
-              "name": "distance",
-              "fee": 8000,
-              "message": ""
-            },
-            {
-              "name": "time",
-              "fee": 0,
-              "message": ""
-            },
-            {
-              "name": "special",
-              "fee": 0,
-              "message": ""
-            }
-          ],
-          "totalFees": 8000,
-          "message": "",
-          "title": "Delivery Charge",
-          "amount": "8000",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "cid=5744630~p=13~eid=00000185-ee73-559a-119c-e34000830d58",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "9.8 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "648823",
-          "deliveryTime": 40,
-          "minDeliveryTime": 40,
-          "maxDeliveryTime": 40,
-          "lastMileTravel": 9.899999618530273,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": true,
-        "avgRating": "4.1",
-        "totalRatings": 50,
-        "new": true
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "23119",
-        "name": "Pizza Hut",
-        "uuid": "8f31bfef-f263-455e-b09e-fe84d824d50c",
-        "city": "81",
-        "area": "Sector 35",
-        "totalRatingsString": "1000+ ratings",
-        "cloudinaryImageId": "joqmtwgzum20i2ixpbvw",
-        "cuisines": [
-          "Fast Food",
-          "Pizzas"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 35000,
-        "costForTwoString": "₹350 FOR TWO",
-        "deliveryTime": 44,
-        "minDeliveryTime": 40,
-        "maxDeliveryTime": 50,
-        "slaString": "40-50 MINS",
-        "lastMileTravel": 3.5,
-        "slugs": {
-          "restaurant": "pizza-hut-crown-interior-mall-greenfields_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "Shop No 19&20, Crown Interior Mall, Faridabad",
-        "locality": "Crown Interior Mall",
-        "parentId": 721,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "50% off",
-          "shortDescriptionList": [
-            {
-              "meta": "50% off | Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "50% off up to ₹100 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "50% OFF",
-          "shortDescriptionList": [
-            {
-              "meta": "Use TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "50% off up to ₹100 | Use code TRYNEW",
-              "discountType": "Percentage",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "3.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "23119",
-          "deliveryTime": 44,
-          "minDeliveryTime": 40,
-          "maxDeliveryTime": 50,
-          "lastMileTravel": 3.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.6",
-        "totalRatings": 1000,
-        "new": false
-      },
-      "subtype": "basic"
-    },
-    {
-      "type": "restaurant",
-      "data": {
-        "type": "F",
-        "id": "314637",
-        "name": "BOX OF CAKE",
-        "uuid": "5d8a36d1-e17f-467a-b11b-046b411ad85d",
-        "city": "81",
-        "area": "Sector 37",
-        "totalRatingsString": "100+ ratings",
-        "cloudinaryImageId": "wtgazpw2i3m7dhg0uu78",
-        "cuisines": [
-          "Desserts",
-          "Snacks"
-        ],
-        "tags": [
-          
-        ],
-        "costForTwo": 40000,
-        "costForTwoString": "₹400 FOR TWO",
-        "deliveryTime": 31,
-        "minDeliveryTime": 31,
-        "maxDeliveryTime": 31,
-        "slaString": "31 MINS",
-        "lastMileTravel": 1.5,
-        "slugs": {
-          "restaurant": "box-of-cake-greenfields_fbd-greenfields_fbd",
-          "city": "faridabad"
-        },
-        "cityState": "81",
-        "address": "SYED MANZER, SCF-142,  HUDA market, Sec 37, Faridabad 122001",
-        "locality": "Greenfields_FBD",
-        "parentId": 15873,
-        "unserviceable": false,
-        "veg": false,
-        "select": false,
-        "favorite": false,
-        "tradeCampaignHeaders": [
-          
-        ],
-        "aggregatedDiscountInfo": {
-          "header": "FREE DELIVERY",
-          "shortDescriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "aggregatedDiscountInfoV2": {
-          "header": "",
-          "shortDescriptionList": [
-            {
-              "meta": "Free Delivery",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "descriptionList": [
-            {
-              "meta": "FREE DELIVERY",
-              "discountType": "FREE_DELIVERY",
-              "operationType": "RESTAURANT"
-            }
-          ],
-          "subHeader": "",
-          "headerType": 0,
-          "superFreedel": ""
-        },
-        "chain": [
-          
-        ],
-        "feeDetails": {
-          "fees": [
-            
-          ],
-          "totalFees": 0,
-          "message": "",
-          "title": "",
-          "amount": "",
-          "icon": ""
-        },
-        "availability": {
-          "opened": true,
-          "nextOpenMessage": "",
-          "nextCloseMessage": ""
-        },
-        "longDistanceEnabled": 0,
-        "rainMode": "NONE",
-        "thirdPartyAddress": false,
-        "thirdPartyVendor": "",
-        "adTrackingID": "",
-        "badges": {
-          "imageBased": [
-            
-          ],
-          "textBased": [
-            
-          ],
-          "textExtendedBadges": [
-            
-          ]
-        },
-        "lastMileTravelString": "1.5 kms",
-        "hasSurge": false,
-        "sla": {
-          "restaurantId": "314637",
-          "deliveryTime": 31,
-          "minDeliveryTime": 31,
-          "maxDeliveryTime": 31,
-          "lastMileTravel": 1.5,
-          "lastMileDistance": 0,
-          "serviceability": "SERVICEABLE",
-          "rainMode": "NONE",
-          "longDistance": "NOT_LONG_DISTANCE",
-          "preferentialService": false,
-          "iconType": "EMPTY"
-        },
-        "promoted": false,
-        "avgRating": "3.9",
-        "totalRatings": 100,
-        "new": false
-      },
-      "subtype": "basic"
-    }
-  ];
